separates the UI from the Backend in HandView to improve compatibility with various operating systems and devices. The codebase has been refactored to remove any breaking dependencies and comprehensive testing has been performed to ensure that the application still works as expected.
